+++ api/apps/user-service/src/models/rolePermission/rolePermission.route.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import getRolePermissionsHandler from '@apps/user-service/src/models/rolePermission/handlers/getRolePermissions.handler';
 import assignPermissionHandler from '@apps/user-service/src/models/rolePermission/handlers/assignPermission.handler';
-import authMiddleware from '@libs/helpers/middlewares/auth.middleware';
+import authMiddleware from '@libs/middlewares/auth.middleware';
 import removePermissionHandler from '@apps/user-service/src/models/rolePermission/handlers/removePermission.handler';
 
 const router = Router();
